requirements.txt
flask
flask-sqlalchemy
google-genai
google-generativeai
weasyprint
pdfkit
python-dotenv
python-docx